fix: expose Railway usage on package config
